Split form config, code review comments
