+++ src/applications/representative-form-upload/tests/e2e/fixtures/data/veteran.json
@@ -1,11 +1,11 @@
 {
   "data": {
     "veteranFullName": {
-      "first": "Hector",
-      "last": "Allen"
+      "first": "Chad",
+      "last": "Barrett"
     },
-    "veteranSsn": "796126859",
-    "veteranDateOfBirth": "1932-02-05",
+    "veteranSsn": "796263749",
+    "veteranDateOfBirth": "1975-05-12",
     "address": {
       "country": "USA",
       "street": "123 Any St",
